Implement single window mode option in parallel to multi-window support
For debugging and development purposes it's sometimes useful to run the
full Android system in a single and statically sized window. This is not
going to be a widely used mode but can be helpful at times.

Also makes our window/layer management a bit more modular to allow easy
additions of new strategies/impementations.